add new detection via the macos logging facility
There are other tools, such as slack huddles, where we want our on-air
light to turn on. Watching the macOS logs to see when an
AVCaptureSession is active is a more general solution and fits in the
spirit of "the familiy doesn't want to be on camera or be recorded" that
started this little project.

Added a legacy option so that the old method can still be used if
needed.